feat: Profile Image Picking + Skills Selection
