+++ src/components/our-components/accountImage.tsx
@@ -1,17 +1,13 @@
-import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
+import { useUser } from '@/context/UserContext';
 
 const AvatarImage = () => {
   const navigate = useNavigate();
-  const [avatarUrl, setAvatarUrl] = useState<string>('');
+  const { user } = useUser();
 
   const handleClick = () => {
     alert('Go to Account Page to see Your Account');
     navigate('/account');
-
-    // Avoiding no-unused-vars problems, I will remove later... - Best
-    setAvatarUrl('');
-    // ===============================================================
   };
 
   return (
@@ -19,9 +15,9 @@ const AvatarImage = () => {
       className="w-[52px] h-[52px] bg-background-profile rounded-full mx-auto flex items-center justify-center overflow-hidden cursor-pointer"
       onClick={handleClick}
     >
-      {avatarUrl ? (
+      {user.avatarUrl ? (
         <img
-          src={avatarUrl}
+          src={user.avatarUrl}
           alt="Avatar Image"
           className="w-full h-full object-cover"
         />
