+++ src/layouts/MainLayout.tsx
@@ -1,10 +1,11 @@
+import { UserProvider } from '@/context/UserContext';
 import { type ReactNode } from 'react';
 
 const MainLayout = ({ children }: { children: ReactNode }) => {
   return (
     <div className="min-h-screen w-full bg-background font-serif">
       {/* Add Other Layout Here */}
-      {children}
+      <UserProvider>{children}</UserProvider>
     </div>
   );
 };
