--- src/pages/AccountStaticPage.tsx
@@ -1,13 +1,0 @@
-export default function AccountStaticPage() {
-  return (
-    <>
-      {/* Header */}
-      <h1 className="text-2xl font-bold mb-2 opacity-0">
-        This is Static Page for Account
-      </h1>
-
-      {/* Content */}
-      <div className="w-full h-full flex flex-col items-center bg-white border rounded-2xl p-8 shadow-sm m-0"></div>
-    </>
-  );
-}
