+++ src/route.tsx
@@ -1,11 +1,10 @@
 import { createBrowserRouter } from 'react-router-dom';
-import ErrorPage from './pages/ErrorPage';
+import ErrorPage from './pages/Error';
 import AccountLayout from './layouts/AccountLayout';
-import LandingPage from './pages/LandingPage';
-import ProfilePage from './pages/ProfilePage';
-import YourServicePage from './pages/YourServicePage';
-import AccountStaticPage from './pages/AccountStaticPage';
-import LoginPage from './pages/LoginPage';
+import LandingPage from './pages/Landing';
+import YourServicePage from './pages/YourService';
+import LoginPage from './pages/Login';
+import ProfilePage from './pages/Profile';
 
 export const router = createBrowserRouter([
   {
@@ -18,8 +17,7 @@ export const router = createBrowserRouter([
     element: <AccountLayout />,
     errorElement: <ErrorPage />,
     children: [
-      { path: '', element: <AccountStaticPage /> },
-      { path: 'profile', element: <ProfilePage /> },
+      { path: '', element: <ProfilePage /> },
       { path: 'service', element: <YourServicePage /> },
     ],
   },
